requirements.txt
fastapi
ponyorm
pandas
python-multipart
pendulum
bs4
requests
